Refactor tests for SRv6
