+++ test/integration/recover_remote_nse_nsm_heal_test.go
@@ -8,19 +8,16 @@ import (
 	"testing"
 	"time"
 
-	"github.com/networkservicemesh/networkservicemesh/controlplane/pkg/properties"
-
 	"github.com/golang/protobuf/ptypes/empty"
 	. "github.com/onsi/gomega"
 	"github.com/sirupsen/logrus"
 
 	"github.com/networkservicemesh/networkservicemesh/controlplane/api/registry"
-	"github.com/networkservicemesh/networkservicemesh/controlplane/pkg/nsmd"
 	"github.com/networkservicemesh/networkservicemesh/test/kubetest"
 	"github.com/networkservicemesh/networkservicemesh/test/kubetest/pods"
 )
 
-func TestNSMHealRemoteDieNSMD_NSE_VXLAN(t *testing.T) {
+func TestNSMHealRemoteDieNSMD_NSE(t *testing.T) {
 	if testing.Short() {
 		t.Skip("Skip, please run without -short")
 		return
@@ -29,144 +26,13 @@ func TestNSMHealRemoteDieNSMD_NSE_VXLAN(t *testing.T) {
 	testNSMHealRemoteDieNSMD_NSE(t, "VXLAN")
 }
 
-func TestNSMHealRemoteDieNSMD_NSE_SRv6(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skip, please run without -short")
-		return
-	}
-
-	testNSMHealRemoteDieNSMD_NSE(t, "SRV6")
-}
-
-func testNSMHealRemoteDieNSMD_NSE(t *testing.T, remoteMechanism string) {
-	g := NewWithT(t)
-
-	k8s, err := kubetest.NewK8s(g, true)
-	defer k8s.Cleanup()
-
-	g.Expect(err).To(BeNil())
-	defer kubetest.MakeLogsSnapshot(k8s, t)
-	// Deploy open tracing to see what happening.
-	config := []*pods.NSMgrPodConfig{
-		{
-			Variables: map[string]string{
-				properties.NsmdHealDSTWaitTimeout: "20", // 20 second delay, since we know both NSM and NSE will die and we need to go with different code branch.
-				nsmd.NsmdDeleteLocalRegistry:      "true",
-			},
-			Namespace:          k8s.GetK8sNamespace(),
-			ForwarderVariables: kubetest.DefaultForwarderVariables(k8s.GetForwardingPlane()),
-		},
-		{
-			Namespace:          k8s.GetK8sNamespace(),
-			Variables:          pods.DefaultNSMD(),
-			ForwarderVariables: kubetest.DefaultForwarderVariables(k8s.GetForwardingPlane()),
-		},
-	}
-	for i := 0; i < 2; i++ {
-		config[i].Variables[nsmd.NsmdPreferredRemoteMechanism] = remoteMechanism
-	}
-	nodes_setup, err := kubetest.SetupNodesConfig(k8s, 2, defaultTimeout, config, k8s.GetK8sNamespace())
-	g.Expect(err).To(BeNil())
-
-	// Run ICMP on latest node
-	icmpPod := kubetest.DeployICMPWithConfig(k8s, nodes_setup[1].Node, "icmp-responder-nse-1", defaultTimeout, 30)
-
-	nscPodNode := kubetest.DeployNSC(k8s, nodes_setup[0].Node, "nsc-1", defaultTimeout)
-	kubetest.CheckNSC(k8s, nscPodNode)
-
-	logrus.Infof("Delete Remote NSMD/ICMP responder NSE")
-	k8s.DeletePods(nodes_setup[1].Nsmd)
-	k8s.DeletePods(icmpPod)
-	//k8s.DeletePods(nodes_setup[1].Nsmd, icmpPod)
-	logrus.Infof("Waiting for NSE with network service")
-	k8s.WaitLogsContains(nodes_setup[0].Nsmd, "nsmd", "Waiting for NSE with network service icmp-responder", time.Minute)
-	// Now are are in forwarder dead state, and in Heal procedure waiting for forwarder.
-	nsmdName := fmt.Sprintf("nsmd-worker-recovered-%d", 1)
-
-	logrus.Infof("Starting recovered NSMD...")
-	startTime := time.Now()
-	nodes_setup[1].Nsmd = k8s.CreatePod(pods.NSMgrPodWithConfig(nsmdName, nodes_setup[1].Node, &pods.NSMgrPodConfig{Namespace: k8s.GetK8sNamespace()})) // Recovery NSEs
-	// Wait for NSMgr to be deployed, to not get admission error
-	kubetest.WaitNSMgrDeployed(k8s, nodes_setup[1].Nsmd, defaultTimeout)
-	logrus.Printf("Started new NSMD: %v on node %s", time.Since(startTime), nodes_setup[1].Node.Name)
-
-	// Restore ICMP responder pod.
-	icmpPod = kubetest.DeployICMP(k8s, nodes_setup[1].Node, "icmp-responder-nse-2", defaultTimeout)
-
-	logrus.Infof("Waiting for connection recovery...")
-	k8s.WaitLogsContains(nodes_setup[0].Nsmd, "nsmd", "Heal: Connection recovered:", defaultTimeout)
-	logrus.Infof("Waiting for connection recovery Done...")
-
-	kubetest.HealNscChecker(k8s, nscPodNode)
-}
-
-func TestNSMHealRemoteDieNSMDVXLAN(t *testing.T) {
+func TestNSMHealRemoteDieNSMD(t *testing.T) {
 	if testing.Short() {
 		t.Skip("Skip, please run without -short")
 		return
 	}
 
 	testNSMHealRemoteDieNSMD(t, "VXLAN")
-}
-
-func TestNSMHealRemoteDieNSMDSRv6(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skip, please run without -short")
-		return
-	}
-
-	testNSMHealRemoteDieNSMD(t, "SRV6")
-}
-
-func testNSMHealRemoteDieNSMD(t *testing.T, remoteMechanism string) {
-	g := NewWithT(t)
-
-	k8s, err := kubetest.NewK8s(g, true)
-	defer k8s.Cleanup()
-
-	g.Expect(err).To(BeNil())
-
-	// Deploy open tracing to see what happening.
-	config := []*pods.NSMgrPodConfig{}
-	for i := 0; i < 2; i++ {
-		cfg := &pods.NSMgrPodConfig{
-			Namespace:          k8s.GetK8sNamespace(),
-			Variables:          pods.DefaultNSMD(),
-			ForwarderVariables: kubetest.DefaultForwarderVariables(k8s.GetForwardingPlane()),
-		}
-		cfg.Variables[nsmd.NsmdPreferredRemoteMechanism] = remoteMechanism
-		config = append(config, cfg)
-	}
-	nodes_setup, err := kubetest.SetupNodesConfig(k8s, 2, defaultTimeout, config, k8s.GetK8sNamespace())
-	g.Expect(err).To(BeNil())
-
-	// Run ICMP on latest node
-	icmpPod := kubetest.DeployICMP(k8s, nodes_setup[1].Node, "icmp-responder-nse-1", defaultTimeout)
-	g.Expect(icmpPod).ToNot(BeNil())
-
-	nscPodNode := kubetest.DeployNSC(k8s, nodes_setup[0].Node, "nsc-1", defaultTimeout)
-	kubetest.CheckNSC(k8s, nscPodNode)
-
-	logrus.Infof("Delete Remote NSMD")
-	k8s.DeletePods(nodes_setup[1].Nsmd)
-
-	logrus.Infof("Waiting for NSE with network service")
-	k8s.WaitLogsContains(nodes_setup[0].Nsmd, "nsmd", "Waiting for NSE with network service icmp-responder", defaultTimeout)
-	// Now are are in forwarder dead state, and in Heal procedure waiting for forwarder.
-	nsmdName := fmt.Sprintf("nsmd-worker-recovered-%d", 1)
-
-	logrus.Infof("Starting recovered NSMD...")
-	startTime := time.Now()
-	nodes_setup[1].Nsmd = k8s.CreatePod(pods.NSMgrPodWithConfig(nsmdName, nodes_setup[1].Node, &pods.NSMgrPodConfig{Namespace: k8s.GetK8sNamespace()})) // Recovery NSEs
-	// Wait for NSMgr to be deployed, to not get admission error
-	kubetest.WaitNSMgrDeployed(k8s, nodes_setup[1].Nsmd, defaultTimeout)
-	logrus.Printf("Started new NSMD: %v on node %s", time.Since(startTime), nodes_setup[1].Node.Name)
-
-	logrus.Infof("Waiting for connection recovery...")
-	k8s.WaitLogsContains(nodes_setup[0].Nsmd, "nsmd", "Heal: Connection recovered:", defaultTimeout)
-	logrus.Infof("Waiting for connection recovery Done...")
-
-	kubetest.HealNscChecker(k8s, nscPodNode)
 }
 
 func TestNSMHealRemoteDieNSMDFakeEndpoint(t *testing.T) {
