+++ test/integration/usecase_nsc_icmp_configuration.go
@@ -1,0 +1,55 @@
+package nsmd_integration_tests
+
+import (
+	"github.com/networkservicemesh/networkservicemesh/controlplane/pkg/nsmd"
+	"github.com/networkservicemesh/networkservicemesh/test/kubetest"
+	"github.com/networkservicemesh/networkservicemesh/test/kubetest/pods"
+	. "github.com/onsi/gomega"
+	v1 "k8s.io/api/core/v1"
+	"testing"
+)
+
+/**
+If passed 1 both will be on same node, if not on different.
+*/
+func testNSCAndICMP(t *testing.T, nodesCount int, useWebhook bool, disableVHost bool, remoteMechanism string) {
+	g := NewWithT(t)
+
+	k8s, err := kubetest.NewK8s(g, true)
+	defer k8s.Cleanup()
+	defer kubetest.MakeLogsSnapshot(k8s, t)
+	g.Expect(err).To(BeNil())
+
+	if useWebhook {
+		awc, awDeployment, awService := kubetest.DeployAdmissionWebhook(k8s, "nsm-admission-webhook", "networkservicemesh/admission-webhook", k8s.GetK8sNamespace(), defaultTimeout)
+		defer kubetest.DeleteAdmissionWebhook(k8s, "nsm-admission-webhook-certs", awc, awDeployment, awService, k8s.GetK8sNamespace())
+	}
+
+	config := []*pods.NSMgrPodConfig{}
+	for i := 0; i < nodesCount; i++ {
+		cfg := &pods.NSMgrPodConfig{
+			Variables: pods.DefaultNSMD(),
+		}
+		cfg.Variables[nsmd.NsmdPreferredRemoteMechanism] = remoteMechanism
+		cfg.Namespace = k8s.GetK8sNamespace()
+		cfg.ForwarderVariables = kubetest.DefaultForwarderVariables(k8s.GetForwardingPlane())
+		if disableVHost {
+			cfg.ForwarderVariables["FORWARDER_ALLOW_VHOST"] = "false"
+		}
+		config = append(config, cfg)
+	}
+	nodes_setup, err := kubetest.SetupNodesConfig(k8s, nodesCount, defaultTimeout, config, k8s.GetK8sNamespace())
+	g.Expect(err).To(BeNil())
+
+	// Run ICMP on latest node
+	_ = kubetest.DeployICMP(k8s, nodes_setup[nodesCount-1].Node, "icmp-responder-nse-1", defaultTimeout)
+
+	var nscPodNode *v1.Pod
+	if useWebhook {
+		nscPodNode = kubetest.DeployNSCWebhook(k8s, nodes_setup[0].Node, "nsc-1", defaultTimeout)
+	} else {
+		nscPodNode = kubetest.DeployNSC(k8s, nodes_setup[0].Node, "nsc-1", defaultTimeout)
+	}
+
+	kubetest.CheckNSC(k8s, nscPodNode)
+}
