+++ test/integration/usecase_nsc_icmp_configuration_srv6_test.go
@@ -1,37 +1,10 @@
-// +build usecase
+// +build srv6
 
 package nsmd_integration_tests
 
 import (
-	"strings"
 	"testing"
-
-	"github.com/networkservicemesh/networkservicemesh/controlplane/pkg/nsmd"
-
-	. "github.com/onsi/gomega"
-	v1 "k8s.io/api/core/v1"
-
-	"github.com/networkservicemesh/networkservicemesh/test/kubetest"
-	"github.com/networkservicemesh/networkservicemesh/test/kubetest/pods"
 )
-
-func TestNSCAndICMPLocal(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skip, please run without -short")
-		return
-	}
-
-	testNSCAndICMP(t, 1, false, false, "")
-}
-
-func TestNSCAndICMPRemoteVXLAN(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skip, please run without -short")
-		return
-	}
-
-	testNSCAndICMP(t, 2, false, false, "VXLAN")
-}
 
 func TestNSCAndICMPRemoteSRv6(t *testing.T) {
 	if testing.Short() {
@@ -40,24 +13,6 @@ func TestNSCAndICMPRemoteSRv6(t *testing.T) {
 	}
 
 	testNSCAndICMP(t, 2, false, false, "SRV6")
-}
-
-func TestNSCAndICMPWebhookLocal(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skip, please run without -short")
-		return
-	}
-
-	testNSCAndICMP(t, 1, true, false, "")
-}
-
-func TestNSCAndICMPWebhookRemoteVXLAN(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skip, please run without -short")
-		return
-	}
-
-	testNSCAndICMP(t, 2, true, false, "SRV6")
 }
 
 func TestNSCAndICMPWebhookRemoteSRv6(t *testing.T) {
@@ -69,24 +24,6 @@ func TestNSCAndICMPWebhookRemoteSRv6(t *testing.T) {
 	testNSCAndICMP(t, 2, true, false, "SRV6")
 }
 
-func TestNSCAndICMPLocalVeth(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skip, please run without -short")
-		return
-	}
-
-	testNSCAndICMP(t, 1, false, true, "")
-}
-
-func TestNSCAndICMPRemoteVethVXLAN(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skip, please run without -short")
-		return
-	}
-
-	testNSCAndICMP(t, 2, false, true, "VXLAN")
-}
-
 func TestNSCAndICMPRemoteVethSRv6(t *testing.T) {
 	if testing.Short() {
 		t.Skip("Skip, please run without -short")
@@ -94,87 +31,4 @@ func TestNSCAndICMPRemoteVethSRv6(t *testing.T) {
 	}
 
 	testNSCAndICMP(t, 2, false, true, "SRV6")
-}
-
-func TestNSCAndICMPNeighbors(t *testing.T) {
-	g := NewWithT(t)
-
-	if testing.Short() {
-		t.Skip("Skip, please run without -short")
-		return
-	}
-
-	k8s, err := kubetest.NewK8s(g, true)
-	defer k8s.Cleanup()
-	defer kubetest.MakeLogsSnapshot(k8s, t)
-	g.Expect(err).To(BeNil())
-
-	nodes_setup, err := kubetest.SetupNodes(k8s, 1, defaultTimeout)
-	g.Expect(err).To(BeNil())
-	_ = kubetest.DeployNeighborNSE(k8s, nodes_setup[0].Node, "icmp-responder-nse-1", defaultTimeout)
-	nsc := kubetest.DeployNSC(k8s, nodes_setup[0].Node, "nsc-1", defaultTimeout)
-
-	pingCommand := "ping"
-	pingIP := "172.16.1.2"
-	arpCommand := []string{"arp", "-a"}
-	if k8s.UseIPv6() {
-		pingCommand = "ping6"
-		pingIP = "100::2"
-		arpCommand = []string{"ip", "-6", "neigh", "show"}
-	}
-	pingResponse, errOut, err := k8s.Exec(nsc, nsc.Spec.Containers[0].Name, pingCommand, pingIP, "-A", "-c", "5")
-	g.Expect(err).To(BeNil())
-	g.Expect(errOut).To(Equal(""))
-	g.Expect(strings.Contains(pingResponse, "100% packet loss")).To(Equal(false))
-
-	nsc2 := kubetest.DeployNSC(k8s, nodes_setup[0].Node, "nsc-2", defaultTimeout)
-	arpResponse, errOut, err := k8s.Exec(nsc2, nsc.Spec.Containers[0].Name, arpCommand...)
-	g.Expect(err).To(BeNil())
-	g.Expect(errOut).To(Equal(""))
-	g.Expect(strings.Contains(arpResponse, pingIP)).To(Equal(true))
-}
-
-/**
-If passed 1 both will be on same node, if not on different.
-*/
-func testNSCAndICMP(t *testing.T, nodesCount int, useWebhook bool, disableVHost bool, remoteMechanism string) {
-	g := NewWithT(t)
-
-	k8s, err := kubetest.NewK8s(g, true)
-	//defer k8s.Cleanup()
-	//defer kubetest.MakeLogsSnapshot(k8s, t)
-	g.Expect(err).To(BeNil())
-
-	if useWebhook {
-		awc, awDeployment, awService := kubetest.DeployAdmissionWebhook(k8s, "nsm-admission-webhook", "networkservicemesh/admission-webhook", k8s.GetK8sNamespace(), defaultTimeout)
-		defer kubetest.DeleteAdmissionWebhook(k8s, "nsm-admission-webhook-certs", awc, awDeployment, awService, k8s.GetK8sNamespace())
-	}
-
-	config := []*pods.NSMgrPodConfig{}
-	for i := 0; i < nodesCount; i++ {
-		cfg := &pods.NSMgrPodConfig{
-			Variables: pods.DefaultNSMD(),
-		}
-		cfg.Variables[nsmd.NsmdPreferredRemoteMechanism] = remoteMechanism
-		cfg.Namespace = k8s.GetK8sNamespace()
-		cfg.ForwarderVariables = kubetest.DefaultForwarderVariables(k8s.GetForwardingPlane())
-		if disableVHost {
-			cfg.ForwarderVariables["FORWARDER_ALLOW_VHOST"] = "false"
-		}
-		config = append(config, cfg)
-	}
-	nodes_setup, err := kubetest.SetupNodesConfig(k8s, nodesCount, defaultTimeout, config, k8s.GetK8sNamespace())
-	g.Expect(err).To(BeNil())
-
-	// Run ICMP on latest node
-	_ = kubetest.DeployICMP(k8s, nodes_setup[nodesCount-1].Node, "icmp-responder-nse-1", defaultTimeout)
-
-	var nscPodNode *v1.Pod
-	if useWebhook {
-		nscPodNode = kubetest.DeployNSCWebhook(k8s, nodes_setup[0].Node, "nsc-1", defaultTimeout)
-	} else {
-		nscPodNode = kubetest.DeployNSC(k8s, nodes_setup[0].Node, "nsc-1", defaultTimeout)
-	}
-
-	kubetest.CheckNSC(k8s, nscPodNode)
 }
