+++ test/integration/usecase_nsc_icmp_configuration_test.go
@@ -6,13 +6,8 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/networkservicemesh/networkservicemesh/controlplane/pkg/nsmd"
-
-	. "github.com/onsi/gomega"
-	v1 "k8s.io/api/core/v1"
-
 	"github.com/networkservicemesh/networkservicemesh/test/kubetest"
-	"github.com/networkservicemesh/networkservicemesh/test/kubetest/pods"
+	. "github.com/onsi/gomega"
 )
 
 func TestNSCAndICMPLocal(t *testing.T) {
@@ -24,22 +19,13 @@ func TestNSCAndICMPLocal(t *testing.T) {
 	testNSCAndICMP(t, 1, false, false, "")
 }
 
-func TestNSCAndICMPRemoteVXLAN(t *testing.T) {
+func TestNSCAndICMPRemote(t *testing.T) {
 	if testing.Short() {
 		t.Skip("Skip, please run without -short")
 		return
 	}
 
 	testNSCAndICMP(t, 2, false, false, "VXLAN")
-}
-
-func TestNSCAndICMPRemoteSRv6(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skip, please run without -short")
-		return
-	}
-
-	testNSCAndICMP(t, 2, false, false, "SRV6")
 }
 
 func TestNSCAndICMPWebhookLocal(t *testing.T) {
@@ -51,16 +37,7 @@ func TestNSCAndICMPWebhookLocal(t *testing.T) {
 	testNSCAndICMP(t, 1, true, false, "")
 }
 
-func TestNSCAndICMPWebhookRemoteVXLAN(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skip, please run without -short")
-		return
-	}
-
-	testNSCAndICMP(t, 2, true, false, "SRV6")
-}
-
-func TestNSCAndICMPWebhookRemoteSRv6(t *testing.T) {
+func TestNSCAndICMPWebhookRemote(t *testing.T) {
 	if testing.Short() {
 		t.Skip("Skip, please run without -short")
 		return
@@ -78,22 +55,13 @@ func TestNSCAndICMPLocalVeth(t *testing.T) {
 	testNSCAndICMP(t, 1, false, true, "")
 }
 
-func TestNSCAndICMPRemoteVethVXLAN(t *testing.T) {
+func TestNSCAndICMPRemoteVeth(t *testing.T) {
 	if testing.Short() {
 		t.Skip("Skip, please run without -short")
 		return
 	}
 
 	testNSCAndICMP(t, 2, false, true, "VXLAN")
-}
-
-func TestNSCAndICMPRemoteVethSRv6(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skip, please run without -short")
-		return
-	}
-
-	testNSCAndICMP(t, 2, false, true, "SRV6")
 }
 
 func TestNSCAndICMPNeighbors(t *testing.T) {
@@ -132,49 +100,4 @@ func TestNSCAndICMPNeighbors(t *testing.T) {
 	g.Expect(err).To(BeNil())
 	g.Expect(errOut).To(Equal(""))
 	g.Expect(strings.Contains(arpResponse, pingIP)).To(Equal(true))
-}
-
-/**
-If passed 1 both will be on same node, if not on different.
-*/
-func testNSCAndICMP(t *testing.T, nodesCount int, useWebhook bool, disableVHost bool, remoteMechanism string) {
-	g := NewWithT(t)
-
-	k8s, err := kubetest.NewK8s(g, true)
-	defer k8s.Cleanup()
-	defer kubetest.MakeLogsSnapshot(k8s, t)
-	g.Expect(err).To(BeNil())
-
-	if useWebhook {
-		awc, awDeployment, awService := kubetest.DeployAdmissionWebhook(k8s, "nsm-admission-webhook", "networkservicemesh/admission-webhook", k8s.GetK8sNamespace(), defaultTimeout)
-		defer kubetest.DeleteAdmissionWebhook(k8s, "nsm-admission-webhook-certs", awc, awDeployment, awService, k8s.GetK8sNamespace())
-	}
-
-	config := []*pods.NSMgrPodConfig{}
-	for i := 0; i < nodesCount; i++ {
-		cfg := &pods.NSMgrPodConfig{
-			Variables: pods.DefaultNSMD(),
-		}
-		cfg.Variables[nsmd.NsmdPreferredRemoteMechanism] = remoteMechanism
-		cfg.Namespace = k8s.GetK8sNamespace()
-		cfg.ForwarderVariables = kubetest.DefaultForwarderVariables(k8s.GetForwardingPlane())
-		if disableVHost {
-			cfg.ForwarderVariables["FORWARDER_ALLOW_VHOST"] = "false"
-		}
-		config = append(config, cfg)
-	}
-	nodes_setup, err := kubetest.SetupNodesConfig(k8s, nodesCount, defaultTimeout, config, k8s.GetK8sNamespace())
-	g.Expect(err).To(BeNil())
-
-	// Run ICMP on latest node
-	_ = kubetest.DeployICMP(k8s, nodes_setup[nodesCount-1].Node, "icmp-responder-nse-1", defaultTimeout)
-
-	var nscPodNode *v1.Pod
-	if useWebhook {
-		nscPodNode = kubetest.DeployNSCWebhook(k8s, nodes_setup[0].Node, "nsc-1", defaultTimeout)
-	} else {
-		nscPodNode = kubetest.DeployNSC(k8s, nodes_setup[0].Node, "nsc-1", defaultTimeout)
-	}
-
-	kubetest.CheckNSC(k8s, nscPodNode)
 }
